requirements.txt
streamlit
tensorflow
streamlit-drawable-canvas
pillow
numpy
opencv-python-headless